Adds linting and pre-commit checks
This updates this repository to use a simular eslint configuration to https://github.com/codeforlansing/cityzen-server, with some minor modifications to support vue.

Additionally, it uses [husky](https://github.com/typicode/husky) to add a pre-commit hook, which run before any commits are made and which ensure that code that does not match our style standards is never commited to the repository.

I have chosen [standardjs](https://standardjs.com/) as our coding standard for these projects, for no particular reason except that it is easy to setup.

Additionally, this includes the result of running the new linter on the files and fixing all the issues it found.

I recommend setting up our editor to run these linters as you type, which will make it less likely that you end up fighting with the linter when you try to commit.

Closes #2