Allows the user to autolaunch the cityzen-server
This resolves the issues with the previous commits. Now npx will
autoinstall the codeforlansing/cityzen server package from github. By
default, it will always pull the latest version on the master branch,
but this can be configured by setting the

`CITYZEN_SERVER_AUTOLAUNCH_BRANCH`

environment variable which which can be used to pull any branch on the github repository, or
by setting the

`CITYZEN_SERVER_AUTOLAUNCH_URL`

environment variable with allows you to specify a custom repository or
even a local directory, instead of the project github.

closes #7